feat: 13 - improve exchange rate table performance
